requirements.txt
Django==4.2.7
djangorestframework==3.14.0
qrcode==7.4.2
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
geoip2==4.7.0
maxminddb==2.6.0
pandas==2.1.3
matplotlib==3.8.2
cairosvg==2.7.1
reportlab==4.0.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
